Module type_list::het_list

A heterogeneous list

Structs

Cons
The heterogeneous list with a head of type T and a tail of type Ts

Nil
The empty heterogeneous list

Traits

AppendItem
A function-trait. Its Output type is the list that represents [..., T] where ... are the types of the elements in Self and its append function provides a means to create a value of that type

HetList
The core trait that any heterogeneous list implements

Functions

cons
a helper function to ease the creation of new heterogeneous lists