Trait type_list::het_list::AppendItem

pub trait AppendItem<T>: HetList {
    type Output: HetList;
    fn append(self, t: T) -> Self::Output;
}

A function-trait. Its Output type is the list that represents [..., T] where ... are the types of the elements in Self and its append function provides a means to create a value of that type

Associated Types

type Output: HetList

The list that represents [..., T] where ... are all elements in Self

Loading content...
Required methods

fn append(self, t: T) -> Self::Output

appends an element to the end of the list

Loading content...