Module type_list::type_list

A list consisting purely of types

Structs

Cons
Cons<T, ...> represents the list [T, ...] where ... are all elements in Ts

Enums

Nil
The empty type list

Traits

AppendType
A function-trait. Its Output type is the list that represents [..., T] where ... are all elements in Self

TypeList
The core trait that any type list implements

Type Definitions

Append
An alias to append a type to the end of a type list